Address review: hedera probe lives with wallet activation, not config
isHederaNetwork exists to decide whether recipients need first-funding
activation — it moves from config.ts to gas-station/wallet-activation.ts,
next to the WalletActivator it gates (config.ts stays pure app-config).
With config.ts out of the test import graph, the .graphql moduleNameMapper
added to jest.operation-context.config.js is unnecessary — reverted.
Also separate the BusinessContract type from the import block in helpers.
